updated nutrient and food data types from integer to float

+++ db/migrate/20200518230631_create_nutrientprofiles.rb
@@ -4,17 +4,17 @@ class CreateNutrientprofiles < ActiveRecord::Migration[6.0]
       t.string :name
       t.decimal :serving
       t.string :size
-      t.integer :weight
-      t.integer :calories, default: 0
-      t.integer :fat, default: 0
-      t.integer :sodium, default: 0
-      t.integer :potassium, default: 0
-      t.integer :carbohydrates, default: 0
-      t.integer :protein, default: 0
-      t.integer :vitaminA, default: 0
-      t.integer :vitaminC, default: 0
-      t.integer :calcium, default: 0
-      t.integer :iron, default: 0
+      t.float :weight
+      t.float :calories, default: 0
+      t.float :fat, default: 0
+      t.float :sodium, default: 0
+      t.float :potassium, default: 0
+      t.float :carbohydrates, default: 0
+      t.float :protein, default: 0
+      t.float :vitaminA, default: 0
+      t.float :vitaminC, default: 0
+      t.float :calcium, default: 0
+      t.float :iron, default: 0
 
       t.timestamps
     end
